+++ src/executor.rs
@@ -734,6 +734,7 @@ fn validate_test_file(
     global_variables: &[test::Variable],
     project: Option<String>,
     environment: Option<String>,
+    index: usize,
 ) -> Option<test::Definition> {
     println!(
         "Variables: {:?}\nVariables2:{:?}",
@@ -743,7 +744,7 @@ fn validate_test_file(
         .name
         .clone()
         .unwrap_or_else(|| test_file.filename.clone());
-    let res = validation::validate_file(test_file, global_variables, project, environment);
+    let res = validation::validate_file(test_file, global_variables, project, environment, index);
     match res {
         Ok(file) => Some(file),
         Err(e) => {
@@ -793,10 +794,10 @@ fn ignored_due_to_tag_filter(
 }
 
 fn schedule_impl(
-    graph: &BTreeMap<String, BTreeSet<String>>,
-    scheduled_nodes: &BTreeSet<String>,
-) -> BTreeSet<String> {
-    let mut ignore: BTreeSet<String> = BTreeSet::new();
+    graph: &BTreeMap<usize, BTreeSet<usize>>,
+    scheduled_nodes: &BTreeSet<usize>,
+) -> BTreeSet<usize> {
+    let mut ignore: BTreeSet<usize> = BTreeSet::new();
     ignore.clone_from(scheduled_nodes);
 
     //Is there a way to do in 1 iteration?
@@ -804,7 +805,7 @@ fn schedule_impl(
         .iter()
         .filter(|(node, _)| !scheduled_nodes.contains(*node))
         .for_each(|(_, edges)| {
-            edges.iter().for_each(|e| _ = ignore.insert(e.clone()));
+            edges.iter().for_each(|e| _ = ignore.insert(*e));
         });
     return graph
         .keys()
@@ -821,56 +822,61 @@ fn construct_test_execution_graph_v2(
         .clone()
         .into_iter()
         .chain(tests_to_ignore)
-        .map(|td| (td.id.clone(), td))
+        .filter(|td| td.id.is_some())
+        .map(|td| (td.id.clone().unwrap(), td))
         .collect();
 
     trace!("determine test execution order based on dependency graph");
 
     //Nodes are IDs ; Directed edges imply ordering; i.e. A -> B; B depends on A
-    let mut graph: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
+    let mut graph: BTreeMap<usize, BTreeSet<usize>> = BTreeMap::new();
     tests_to_run
         .iter()
         .map(|td| (td.id.clone(), td))
         .for_each(|(id, definition)| {
-            if let Some(req) = definition.requires.as_ref() {
-                let required_def = tests_by_id.get(req);
-                if required_def.is_none() {
+            if let Some(required_id) = definition.requires.as_ref() {
+                let Some(required_def) = tests_by_id.get(required_id) else {
                     return;
-                }
+                };
 
-                if required_def.unwrap().disabled {
+                if required_def.disabled {
                     warn!(
                         "Test \"{}\" requires a disabled test: \"{}\"",
-                        definition.name.as_deref().unwrap_or(id.as_str()),
-                        required_def.unwrap().id
+                        definition
+                            .name
+                            .as_deref()
+                            .unwrap_or(&id.clone().unwrap_or(definition.index.to_string())),
+                        required_id
                     );
                     //should we do transitive disablement?
                 }
 
-                if let Some(edges) = graph.get_mut(req) {
-                    edges.insert(id.clone());
+                if let Some(edges) = graph.get_mut(&required_def.index) {
+                    edges.insert(definition.index);
                 } else {
-                    graph.insert(req.clone(), BTreeSet::from([id.clone()]));
+                    graph.insert(required_def.index, BTreeSet::from([definition.index]));
                 }
             }
 
-            let node_for_id = graph.get(&id);
-            if node_for_id.is_none() {
-                graph.insert(id.clone(), BTreeSet::new());
+            let node_for_index = graph.get(&definition.index);
+            if node_for_index.is_none() {
+                graph.insert(definition.index, BTreeSet::new());
             }
             //intution: if it already has a dependent, its simply a test
             //depended on by multiple other tests and not a duplicate ID made in error
-            else if node_for_id.unwrap().is_empty() {
-                warn!("Skipping test, found duplicate test id: {}", id.clone());
+            else if node_for_index.unwrap().is_empty() {
+                warn!(
+                    "Skipping test, found duplicate test id: {}",
+                    &id.clone().unwrap()
+                );
             }
         });
 
-    let mut jobs: Vec<BTreeSet<String>> = Vec::new();
-    let mut scheduled_nodes: BTreeSet<String> = BTreeSet::new();
+    let mut jobs: Vec<BTreeSet<usize>> = Vec::new();
+    let mut scheduled_nodes: BTreeSet<usize> = BTreeSet::new();
     while graph.len() != scheduled_nodes.len() {
         let job = schedule_impl(&graph, &scheduled_nodes);
-        job.iter()
-            .for_each(|n| _ = scheduled_nodes.insert(n.clone()));
+        job.iter().for_each(|n| _ = scheduled_nodes.insert(*n));
         jobs.push(job);
     }
 
@@ -878,7 +884,7 @@ fn construct_test_execution_graph_v2(
         .into_iter()
         .map(|hs| {
             hs.into_iter()
-                .map(|id| tests_by_id.get(&id).unwrap().clone())
+                .map(|index| tests_to_run.get(index).unwrap().clone())
                 .collect::<Vec<Definition>>()
         })
         .collect();
@@ -892,11 +898,19 @@ fn construct_test_execution_graph_v2(
         //not smart enough on rust to write generic lambda in order to not repeat myself here
         let s1: HashSet<String> = tests_to_run
             .iter()
-            .map(|td| td.name.clone().unwrap_or(td.id.clone()))
+            .map(|td| {
+                td.name
+                    .clone()
+                    .unwrap_or(td.id.clone().unwrap_or(td.index.to_string()))
+            })
             .collect();
         let s2: HashSet<String> = flattened_jobs
             .iter()
-            .map(|td| td.name.clone().unwrap_or(td.id.clone()))
+            .map(|td| {
+                td.name
+                    .clone()
+                    .unwrap_or(td.id.clone().unwrap_or(td.index.to_string()))
+            })
             .collect();
         let missing_tests = (&s1 - &s2)
             .into_iter()
@@ -916,12 +930,15 @@ fn construct_test_execution_graph_v2(
 
     for (count, job) in job_definitions.iter().enumerate() {
         trace!(
-            "Job {count}, Tests: {}",
-            job.iter().fold("".to_string(), |acc, x| format!(
-                "{},{}",
-                acc,
-                x.name.as_ref().unwrap_or(&x.id)
-            ))
+            "Job {count}: {}",
+            job.iter()
+                .enumerate()
+                .map(|(i, t)| t
+                    .name
+                    .clone()
+                    .unwrap_or(t.id.clone().unwrap_or(i.to_string())))
+                .collect::<Vec<String>>()
+                .join(", ")
         )
     }
 
@@ -941,8 +958,15 @@ pub fn tests_from_files(
     let tests_to_run: Vec<test::Definition> = files
         .into_iter()
         .filter_map(|s| load_test_from_path(s.as_str()))
-        .filter_map(|f| {
-            validate_test_file(f, &global_variables, project.clone(), environment.clone())
+        .enumerate()
+        .filter_map(|(i, f)| {
+            validate_test_file(
+                f,
+                &global_variables,
+                project.clone(),
+                environment.clone(),
+                i,
+            )
         })
         .filter_map(|f| {
             if !ignored_due_to_tag_filter(&f, &tags, &tag_mode) {
@@ -954,6 +978,31 @@ pub fn tests_from_files(
         })
         .collect();
     (tests_to_run, tests_to_ignore)
+}
+
+pub fn print_validation_failures(
+    failures: Vec<(&Definition, telemetry::PlatformIdFailure)>,
+    as_errors: bool,
+) {
+    let messages: Vec<String> = failures.iter().map(|(definition, failure)| {
+        match failure {
+            telemetry::PlatformIdFailure::Missing => {
+                format!("Test at path \"{}\" does not have a platformId. This field is required to stream telemetry data.", definition.file_data.filename)
+            },
+            telemetry::PlatformIdFailure::Invalid => {
+                format!("Test at path \"{}\" has an invalid platformId ({}). This field must be a valid ULID.", definition.file_data.filename, definition.platform_id.clone().expect("PlatformId should have a value"))
+            },
+            telemetry::PlatformIdFailure::Duplicate => {
+                format!("Test at path \"{}\" has a duplicate platformId ({}). This field must be a valid ULID and unique across all files.", definition.file_data.filename, definition.platform_id.clone().expect("PlatformId should have a value"))
+            },
+        }
+   }).collect();
+
+    if as_errors {
+        messages.iter().for_each(|m| error!("{}", m));
+    } else {
+        messages.iter().for_each(|m| warn!("{}", m));
+    }
 }
 
 pub async fn execute_tests(
@@ -976,20 +1025,33 @@ pub async fn execute_tests(
 
     let mut session: Option<telemetry::Session> = None;
 
-    if !mode_dryrun {
-        if let Some(token) = &config.settings.api_key {
-            if let Ok(t) = uuid::Uuid::parse_str(token) {
-                match telemetry::create_session(t, all_tests, cli_args, &config).await {
-                    Ok(sess) => {
-                        session = Some(sess);
+    if let Some(token) = &config.settings.api_key {
+        if let Ok(t) = uuid::Uuid::parse_str(token) {
+            let validation_results =
+                telemetry::validate_platform_ids(tests_to_run.iter().collect());
+
+            if !mode_dryrun {
+                match validation_results {
+                    Ok(_) => {
+                        match telemetry::create_session(t, all_tests, cli_args, &config).await {
+                            Ok(sess) => {
+                                session = Some(sess);
+                            }
+                            Err(e) => {
+                                debug!("telemetry failed: {}", e);
+                            }
+                        }
                     }
-                    Err(e) => {
-                        debug!("telemetry failed: {}", e);
+                    Err(failures) => {
+                        print_validation_failures(failures, true);
+                        return Report::default();
                     }
                 }
-            } else {
-                debug!("invalid api token: {}", &token);
+            } else if let Err(failures) = validation_results {
+                print_validation_failures(failures, false);
             }
+        } else {
+            debug!("invalid api token: {}", &token);
         }
     }
 
@@ -2177,6 +2239,7 @@ mod tests {
     use self::test::definition::ResolvedRequest;
     use hyper::Response;
     use std::any::Any;
+    use test::File;
 
     use super::*;
     use adjacent_pair_iterator::AdjacentPairIterator;
@@ -2542,11 +2605,13 @@ mod tests {
     fn construct_definition_for_dependency_graph(
         id: &str,
         requires: Option<String>,
+        index: usize,
     ) -> test::Definition {
         test::Definition {
             name: None,
             description: None,
-            id: String::from(id),
+            id: Some(id.to_string()),
+            platform_id: None,
             project: None,
             environment: None,
             requires,
@@ -2562,7 +2627,8 @@ mod tests {
                 always: None,
             },
             disabled: false,
-            filename: "/a/path.jkt".to_string(),
+            file_data: File::default(),
+            index,
         }
     }
 
@@ -2570,12 +2636,13 @@ mod tests {
     fn no_dependencies_is_one_execution_node() {
         let defs = vec!["A", "B", "C", "D"]
             .into_iter()
-            .map(|id| construct_definition_for_dependency_graph(id, None))
+            .enumerate()
+            .map(|(i, id)| construct_definition_for_dependency_graph(id, None, i))
             .collect();
 
         let actual = construct_test_execution_graph_v2(
             defs,
-            vec![construct_definition_for_dependency_graph("E", None)],
+            vec![construct_definition_for_dependency_graph("E", None, 4)],
         );
         assert_eq!(1, actual.len());
         assert_eq!(4, actual.get(0).unwrap().len());
@@ -2585,19 +2652,25 @@ mod tests {
     fn one_root_dependency_is_two_execution_nodes() {
         let mut defs = vec!["A", "B", "C", "D"]
             .into_iter()
-            .map(|id| construct_definition_for_dependency_graph(id, Some("Parent".to_string())))
+            .enumerate()
+            .map(|(i, id)| {
+                construct_definition_for_dependency_graph(id, Some("Parent".to_string()), i)
+            })
             .collect::<Vec<Definition>>();
 
-        defs.push(construct_definition_for_dependency_graph("Parent", None));
+        defs.push(construct_definition_for_dependency_graph("Parent", None, 4));
 
         let actual = construct_test_execution_graph_v2(
             defs,
-            vec![construct_definition_for_dependency_graph("E", None)],
+            vec![construct_definition_for_dependency_graph("E", None, 5)],
         );
 
         assert_eq!(2, actual.len());
         assert_eq!(1, actual.get(0).unwrap().len());
-        assert_eq!("Parent", actual.get(0).unwrap().get(0).unwrap().id);
+        assert_eq!(
+            "Parent",
+            actual.get(0).unwrap().get(0).unwrap().id.clone().unwrap()
+        );
         assert_eq!(4, actual.get(1).unwrap().len());
     }
 
@@ -2610,12 +2683,13 @@ mod tests {
             .map(|(pos, (fst, snd))| {
                 let mut res: Vec<Definition> = Vec::new();
                 if pos == 0 {
-                    res.push(construct_definition_for_dependency_graph(fst, None));
+                    res.push(construct_definition_for_dependency_graph(fst, None, 0));
                 }
 
                 res.push(construct_definition_for_dependency_graph(
                     snd,
                     Some(fst.to_string()),
+                    pos + 1,
                 ));
 
                 return res;
@@ -2632,7 +2706,8 @@ mod tests {
         test::Definition {
             name: None,
             description: None,
-            id: String::from("id"),
+            id: None,
+            platform_id: None,
             project: None,
             environment: None,
             requires: None,
@@ -2648,7 +2723,8 @@ mod tests {
                 always: None,
             },
             disabled: false,
-            filename: "/a/path.jkt".to_string(),
+            file_data: File::default(),
+            index: 0,
         }
     }