+++ src/main.rs
@@ -18,7 +18,10 @@ use logger::SimpleLogger;
 use serde::{Deserialize, Serialize};
 use std::error::Error;
 use std::path::{Path, PathBuf};
+use telemetry::PlatformIdFailure;
 use tokio::fs;
+use tokio::io::AsyncWriteExt;
+use ulid::Ulid;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 const IGNORE_FILE: &str = ".jikkenignore";
@@ -28,6 +31,8 @@ pub enum ExecutionMode {
     Run,
     Dryrun,
     List,
+    Format,
+    Validate(bool),
 }
 
 pub enum TagMode {
@@ -74,6 +79,7 @@ pub struct Cli {
 #[derive(Subcommand, Serialize, Deserialize)]
 pub enum Commands {
     /// Execute tests
+    #[command(name = "run", alias = "r")]
     Run {
         /// The path(s) to search for test files
         /// {n}By default, the current path is used
@@ -98,8 +104,8 @@ pub enum Commands {
         junit: Option<String>,
     },
 
-    /// Process tests without calling API endpoints
-    #[command(name = "dryrun")]
+    /// Execute tests without calling API endpoints
+    #[command(name = "dryrun", alias = "dr")]
     DryRun {
         /// The path(s) to search for test files
         /// {n}By default, the current path is used
@@ -124,7 +130,7 @@ pub enum Commands {
         junit: Option<String>,
     },
 
-    /// Lists tests at the given path(s)
+    /// List test files
     #[command(name = "list")]
     List {
         /// The path(s) to search for test files
@@ -144,6 +150,53 @@ pub enum Commands {
         /// Toggle tag matching logic to select tests matching any of the given tags
         #[arg(long, default_value_t = false)]
         tags_or: bool,
+    },
+
+    /// Format test files
+    #[command(name = "format", alias = "fmt")]
+    Format {
+        /// The path(s) to search for test files
+        /// {n}By default, the current path is used
+        #[arg(name = "path")]
+        paths: Vec<String>,
+
+        /// Recursively search for test files
+        #[arg(short)]
+        recursive: bool,
+
+        /// Select tests to format based on tags
+        /// {n}By default, tests must match all given tags to be selected
+        #[arg(short, long = "tag", name = "tag")]
+        tags: Vec<String>,
+
+        /// Toggle tag matching logic to select tests matching any of the given tags
+        #[arg(long, default_value_t = false)]
+        tags_or: bool,
+    },
+
+    /// Validate test files
+    Validate {
+        /// The path(s) to search for test files
+        /// {n}By default, the current path is used
+        #[arg(name = "path")]
+        paths: Vec<String>,
+
+        /// Recursively search for test files
+        #[arg(short)]
+        recursive: bool,
+
+        /// Select tests to validate based on tags
+        /// {n}By default, tests must match all given tags to be selected
+        #[arg(short, long = "tag", name = "tag")]
+        tags: Vec<String>,
+
+        /// Toggle tag matching logic to select tests matching any of the given tags
+        #[arg(long, default_value_t = false)]
+        tags_or: bool,
+
+        /// Automatically generate and insert platform IDs in tests that don't have one
+        #[arg(long, default_value_t = false)]
+        generate_platform_ids: bool,
     },
 
     /// Create a new test
@@ -168,7 +221,7 @@ pub enum Commands {
         output: bool,
     },
 
-    /// Update Jikken, if a newer version exists
+    /// Update Jikken (if a newer version exists)
     Update,
 }
 
@@ -244,7 +297,7 @@ async fn search_directory(
     let entry_is_file = |e: &walkdir::DirEntry| e.metadata().map(|e| e.is_file()).unwrap_or(false);
 
     walkdir::WalkDir::new(path)
-        .max_depth(if recursive { ::std::usize::MAX } else { 1 })
+        .max_depth(if recursive { usize::MAX } else { 1 })
         .into_iter()
         .filter_entry(create_top_level_filter(&glob_pattern))
         .filter_map(|e| e.ok())
@@ -336,19 +389,20 @@ async fn get_files(
     Ok(results)
 }
 
-fn print_test_info(tests: Vec<test::Definition>) {
+fn print_test_info(mut tests: Vec<test::Definition>) {
     let mut path_column = vec!["PATH".to_string()];
     let mut name_column = vec!["TEST NAME".to_string()];
     let mut tags_column = vec!["TAGS".to_string()];
 
+    tests.sort_by_key(|td| td.file_data.filename.clone());
     tests.into_iter().for_each(|td| {
         name_column.push(td.name.unwrap_or("<none>".to_string()));
         tags_column.push(if td.tags.is_empty() {
             "<none>".to_string()
         } else {
-            td.tags.join(",")
+            td.tags.join(", ")
         });
-        path_column.push(td.filename)
+        path_column.push(td.file_data.filename)
     });
 
     let get_column_width = |v: &Vec<String>| v.iter().fold(0, |max, s| std::cmp::max(max, s.len()));
@@ -421,6 +475,62 @@ async fn run_tests(
                 plurality_policy(number_of_tests_to_run)
             );
         }
+        return Ok(executor::Report::default());
+    }
+
+    if execution_mode == ExecutionMode::Format {
+        for td in &tests_to_run {
+            let mut file = fs::File::create(&td.file_data.filename).await?;
+            let file_data = serde_yaml::to_string(&td.file_data).unwrap();
+            file.write_all(file_data.as_bytes()).await?;
+        }
+
+        info!(
+            "Successfully formatted {} test files.\n",
+            &tests_to_run.len()
+        );
+        return Ok(executor::Report::default());
+    }
+
+    if let ExecutionMode::Validate(generate) = execution_mode {
+        if let Some(token) = &config.settings.api_key {
+            if uuid::Uuid::parse_str(token).is_ok() {
+                let validation_results =
+                    telemetry::validate_platform_ids(tests_to_run.iter().collect());
+                let mut has_missing = false;
+
+                if let Err(failures) = validation_results {
+                    for failure in failures {
+                        if generate && failure.1 == PlatformIdFailure::Missing {
+                            let platform_id = Ulid::new().to_string();
+                            let mut test = tests_to_run[failure.0.index].clone();
+                            test.file_data.platform_id = Some(platform_id.clone());
+
+                            let mut file = fs::File::create(&test.file_data.filename).await?;
+                            let file_data = serde_yaml::to_string(&test.file_data).unwrap();
+                            file.write_all(file_data.as_bytes()).await?;
+
+                            info!(
+                                "Successfully updated test at path \"{}\" with platform ID {}.\n",
+                                test.file_data.filename, platform_id
+                            );
+                        } else {
+                            if !has_missing && failure.1 == PlatformIdFailure::Missing {
+                                has_missing = true;
+                            }
+                            executor::print_validation_failures(vec![failure], true);
+                        }
+                    }
+
+                    if has_missing {
+                        warn!("\nRun the validate command with the --generate-platform-ids option to automatically generate and insert missing platform IDs.");
+                    }
+                }
+            } else {
+                error!("API Key {} is invalid", &token);
+            }
+        }
+
         return Ok(executor::Report::default());
     }
 
@@ -532,28 +642,31 @@ async fn main() -> std::process::ExitCode {
     let cli_environment = cli.environment;
 
     let exit_code: std::process::ExitCode = match cli.command {
-        Commands::Update => {
-            updater::try_updating().await;
-            std::process::ExitCode::SUCCESS
-        }
-        Commands::New {
-            full,
-            openapi_spec_path,
-            multistage,
-            output,
-            name,
+        Commands::Run {
+            tags,
+            tags_or,
+            recursive,
+            paths,
+            junit,
         } => {
             updater::check_for_updates().await;
-            match openapi_spec_path {
-                Some(path) => result_to_exit_code(
-                    new::create_tests_from_openapi_spec(path.as_str(), full, multistage, name),
-                    false,
-                ),
-                None => result_to_exit_code(
-                    new::create_test_template(full, multistage, output, name).await,
-                    false,
-                ),
-            }
+            log::logger().flush();
+            check_supplied_config_file_existence(&cli.config_file);
+            result_report_to_exit_code(
+                run_tests(
+                    paths,
+                    tags,
+                    tags_or,
+                    ExecutionMode::Run,
+                    recursive,
+                    cli_project,
+                    cli_environment,
+                    cli.config_file,
+                    junit,
+                    cli_args,
+                )
+                .await,
+            )
         }
         Commands::DryRun {
             tags,
@@ -583,32 +696,6 @@ async fn main() -> std::process::ExitCode {
                 .await,
             )
         }
-        Commands::Run {
-            tags,
-            tags_or,
-            recursive,
-            paths,
-            junit,
-        } => {
-            updater::check_for_updates().await;
-            log::logger().flush();
-            check_supplied_config_file_existence(&cli.config_file);
-            result_report_to_exit_code(
-                run_tests(
-                    paths,
-                    tags,
-                    tags_or,
-                    ExecutionMode::Run,
-                    recursive,
-                    cli_project,
-                    cli_environment,
-                    cli.config_file,
-                    junit,
-                    cli_args,
-                )
-                .await,
-            )
-        }
         Commands::List {
             tags,
             tags_or,
@@ -633,10 +720,82 @@ async fn main() -> std::process::ExitCode {
                 .await,
             )
         }
+        Commands::Format {
+            tags,
+            tags_or,
+            recursive,
+            paths,
+        } => {
+            updater::check_for_updates().await;
+            check_supplied_config_file_existence(&cli.config_file);
+            result_report_to_exit_code(
+                run_tests(
+                    paths,
+                    tags,
+                    tags_or,
+                    ExecutionMode::Format,
+                    recursive,
+                    cli_project,
+                    cli_environment,
+                    cli.config_file,
+                    None,
+                    cli_args,
+                )
+                .await,
+            )
+        }
+        Commands::Validate {
+            tags,
+            tags_or,
+            recursive,
+            paths,
+            generate_platform_ids,
+        } => {
+            updater::check_for_updates().await;
+            check_supplied_config_file_existence(&cli.config_file);
+            result_report_to_exit_code(
+                run_tests(
+                    paths,
+                    tags,
+                    tags_or,
+                    ExecutionMode::Validate(generate_platform_ids),
+                    recursive,
+                    cli_project,
+                    cli_environment,
+                    cli.config_file,
+                    None,
+                    cli_args,
+                )
+                .await,
+            )
+        }
+        Commands::New {
+            full,
+            openapi_spec_path,
+            multistage,
+            output,
+            name,
+        } => {
+            updater::check_for_updates().await;
+            match openapi_spec_path {
+                Some(path) => result_to_exit_code(
+                    new::create_tests_from_openapi_spec(path.as_str(), full, multistage, name),
+                    false,
+                ),
+                None => result_to_exit_code(
+                    new::create_test_template(full, multistage, output, name).await,
+                    false,
+                ),
+            }
+        }
+        Commands::Update => {
+            updater::try_updating().await;
+            std::process::ExitCode::SUCCESS
+        }
     };
 
     log::logger().flush();
-    return exit_code;
+    exit_code
 }
 
 //------------------TESTS---------------------------------