+++ src/test/file.rs
@@ -110,7 +110,7 @@ impl<'de> Deserialize<'de> for VariableName {
 }
 
 //aka RefOrT , where Ref should refer to a variable
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(untagged)]
 pub enum UnvalidatedVariableNameOrComponent<T> {
     VariableName(VariableName),
@@ -244,13 +244,13 @@ impl ValuesOrSchema {
 #[serde(rename_all = "camelCase")]
 pub struct SequenceSpecification {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub schema: Option<ValuesOrSchema>,
-    #[serde(skip_serializing_if = "Option::is_none")]
     pub length: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub min_length: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub max_length: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub schema: Option<ValuesOrSchema>,
 }
 
 #[derive(Hash, Default, Serialize, Debug, Clone, Deserialize, PartialEq)]
@@ -270,12 +270,15 @@ pub struct DateSpecification {
 #[derive(Hash, Default, Serialize, Debug, Clone, Deserialize, PartialEq)]
 pub struct DateTimeSpecification {
     #[serde(flatten)]
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub specification: Option<Specification<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub min: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub max: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub format: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub modifier: Option<variable::Modifier>,
 }
 
@@ -2335,9 +2338,10 @@ impl TryFrom<UnvalidatedDatumSchemaVariable2> for DatumSchema {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct UnvalidatedRequest {
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub method: Option<http::Verb>,
     pub url: String,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -2375,16 +2379,23 @@ impl Hash for UnvalidatedRequest {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct UnvalidatedCompareRequest {
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub method: Option<http::Verb>,
     pub url: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub params: Option<Vec<http::Parameter>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub add_params: Option<Vec<http::Parameter>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub ignore_params: Option<Vec<String>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub headers: Option<Vec<http::Header>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub add_headers: Option<Vec<http::Header>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub ignore_headers: Option<Vec<String>>,
     //Requests can only contain a body OR a body_schema
     //We used to signify this using (serde-flattened)enums, but its
@@ -2393,6 +2404,7 @@ pub struct UnvalidatedCompareRequest {
     //in the (Validated)CompareDescriptor struct
     #[serde(skip_serializing_if = "Option::is_none")]
     pub body: Option<UnvalidatedVariableNameOrValue>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub strict: Option<bool>,
 }
 
@@ -2488,9 +2500,17 @@ pub type UnvalidatedVariableNameOrDatumSchema2 =
 #[derive(Debug, Serialize, Clone, Deserialize, PartialEq)]
 #[serde(untagged)]
 pub enum ValueOrDatumOrFile {
-    File { file: String },
+    File {
+        file: String,
+    },
     Schema(DatumSchema),
-    Value { value: Value },
+    Value {
+        value: Value,
+    },
+    #[serde(rename_all = "camelCase")]
+    ValueSet {
+        value_set: Vec<Value>,
+    },
 }
 
 impl Hash for ValueOrDatumOrFile {
@@ -2500,6 +2520,9 @@ impl Hash for ValueOrDatumOrFile {
             ValueOrDatumOrFile::Schema(s) => s.hash(state),
             ValueOrDatumOrFile::Value { value } => {
                 serde_json::to_string(value).unwrap().hash(state)
+            }
+            ValueOrDatumOrFile::ValueSet { value_set } => {
+                serde_json::to_string(value_set).unwrap().hash(state)
             }
         }
     }
@@ -2638,7 +2661,7 @@ impl<'a> Checker for BodyOrSchemaChecker<'a> {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct UnvalidatedResponse {
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -2919,9 +2942,11 @@ pub struct UnvalidatedVariable {
     pub value: ValueOrDatumOrFile,
 }
 
-#[derive(Hash, Debug, Clone, Serialize, Deserialize)]
+#[derive(Hash, Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields)]
 pub struct UnvalidatedStage {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
     pub request: UnvalidatedRequest,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub compare: Option<UnvalidatedCompareRequest>,
@@ -2932,19 +2957,17 @@ pub struct UnvalidatedStage {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub variables2: Option<Vec<UnvalidatedVariable3>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
     pub delay: Option<u64>,
 }
 
-#[derive(Hash, Debug, Clone, Serialize, Deserialize)]
+#[derive(Hash, Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields)]
 pub struct UnvalidatedRequestResponse {
     pub request: UnvalidatedRequest,
     pub response: Option<UnvalidatedResponse>,
 }
 
-#[derive(Hash, Debug, Clone, Serialize, Deserialize)]
+#[derive(Hash, Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields)]
 pub struct UnvalidatedCleanup {
     pub onsuccess: Option<UnvalidatedRequest>,