+++ src/test/validation.rs
@@ -1,8 +1,12 @@
 use crate::test;
 use crate::test::definition;
+use crate::test::validation;
 use crate::test::variable;
+use log::warn;
+use regex::Regex;
 use std::fmt;
 use std::path::PathBuf;
+use ulid::Ulid;
 
 #[derive(Debug, Clone)]
 pub struct Error {
@@ -22,9 +26,32 @@ impl fmt::Display for Error {
 }
 
 fn validate_test_file(
-    _file: &test::File,
+    file: &test::File,
     _global_variables: &[test::Variable],
 ) -> Result<bool, Error> {
+    if !file
+        .platform_id
+        .clone()
+        .map(|ulid| Ulid::from_string(&ulid).is_ok())
+        .unwrap_or(true)
+    {
+        warn!("Test file ({}) has invalid platform identifier ({}). PlatformId must be empty or a valid ULID.", file.filename, file.platform_id.clone().unwrap_or("".to_string()));
+    }
+
+    let regex = Regex::new(r"(?i)^[a-z0-9-_]+$").unwrap();
+    if !file
+        .id
+        .clone()
+        .map(|id| regex.is_match(id.as_str()))
+        .unwrap_or(true)
+    {
+        return Err(validation::Error {
+            reason:
+                format!("id '{}' is invalid - may only contain alphanumeric characters, hyphens, and underscores", file.id.clone().unwrap())
+                    .to_string(),
+        });
+    }
+
     Ok(true)
 }
 
@@ -34,6 +61,7 @@ pub fn validate_file(
     global_variables: &[test::Variable],
     project: Option<String>,
     environment: Option<String>,
+    index: usize,
 ) -> Result<test::Definition, Error> {
     validate_test_file(&file, global_variables)?;
     let new_tags = if let Some(tags) = file.tags.as_ref() {
@@ -45,10 +73,8 @@ pub fn validate_file(
         Vec::new()
     };
 
-    let generated_id = file.generate_id();
-
     let variables = test::Variable::validate_variables_opt(
-        file.variables,
+        file.clone().variables,
         PathBuf::from(&file.filename)
             .parent()
             .and_then(|p| p.to_str())
@@ -56,7 +82,7 @@ pub fn validate_file(
     )?;
 
     let variables2 = test::Variable::validate_variables_opt2(
-        file.variables2,
+        file.clone().variables2,
         PathBuf::from(&file.filename)
             .parent()
             .and_then(|p| p.to_str())
@@ -64,9 +90,11 @@ pub fn validate_file(
     )?;
 
     let td = test::Definition {
+        file_data: file.clone(),
         name: file.name,
         description: file.description,
-        id: file.id.unwrap_or(generated_id).to_lowercase(),
+        id: file.id.map(|i| i.to_lowercase()),
+        platform_id: file.platform_id,
         project: file.project.or(project),
         environment: file.env.or(environment),
         requires: file.requires,
@@ -86,7 +114,7 @@ pub fn validate_file(
         setup: definition::RequestResponseDescriptor::new_opt(file.setup, &variables)?,
         cleanup: definition::CleanupDescriptor::new(file.cleanup, &variables)?,
         disabled: file.disabled.unwrap_or_default(),
-        filename: file.filename,
+        index,
     };
 
     td.update_variable_matching();